backend/requirements.txt

fastapi
uvicorn
openai
python-dotenv
langchain
chromadb
pydantic
